+++ src/components/modal/post/PostPreviewModal.tsx
@@ -10,8 +10,6 @@ import { ToggleButton, ToggleButtonGroup } from "@mui/material";
 import { useModal } from "hooks/modal/ModalProvider";
 import ModalOption from "enum/modalOptionTypes";
 import PostCardListCarousel from "containers/post/postCardList/web/PostCardLIstCarousel";
-import getMemberInfo from "api/member/getMemberInfo";
-import { IMemberInfoDetails } from "interface/member/IMemberInfoDetails";
 
 interface PostPreviewModalProps {
   postId?: string;
@@ -20,8 +18,6 @@ interface PostPreviewModalProps {
 function PostPreviewModal({ postId }: PostPreviewModalProps) {
   const [postDetails, setPostDetails] = useState<IPostDTO | null>(null);
   const [sideContentType, setSideContentType] = useState<string>("comment");
-  //현재 로그인한 유저 정보 저장
-  const [userInfo, setUserInfo] = useState<IMemberInfoDetails | null>(null);
 
   const { closeModal } = useModal();
   useEffect(() => {
@@ -40,15 +36,6 @@ function PostPreviewModal({ postId }: PostPreviewModalProps) {
       toast.error("로그인이 필요합니다.");
     }
   }, [postId]);
-
-  useEffect(() => {
-    const getUserInfo = async () => {
-      const curMember = await getMemberInfo();
-      setUserInfo(curMember);
-    };
-    getUserInfo();
-    //eslint-disable-next-line
-  }, []);
 
   const handleChange = (
     event: React.MouseEvent<HTMLElement>,
@@ -86,7 +73,6 @@ function PostPreviewModal({ postId }: PostPreviewModalProps) {
               <PostComments
                 postId={postId}
                 memberId={postDetails.memberInfo.memberId}
-                userInfo={userInfo}
               />
             ) : (
               <PostMap cards={postDetails.cards} />
