+++ src/containers/post/mobile/PostDetailsMobile.module.css
